restructured futures class into account.py for ease of access and simplicity
